stateを別に移しただけで何も整理されていない
