+++ src/camera/camera.ts
@@ -1,7 +1,7 @@
+import { getCurrentParams } from "@/mandelbrot-state/mandelbrot-state";
 import type { IterationBuffer } from "@/types";
 import p5 from "p5";
 import { getIterationCache } from "../aggregator/aggregator";
-import { getCurrentParams } from "../mandelbrot";
 import { Rect } from "../rect";
 import { renderIterationsToPixel } from "../rendering/rendering";
 import { getCurrentPalette, markAsRendered, needsRerender } from "./palette";
