+++ src/lib/params.ts
@@ -1,6 +1,6 @@
 import { getCurrentPalette } from "@/camera/palette";
 import { deserializePalette } from "@/color";
-import { getCurrentParams } from "@/mandelbrot";
+import { getCurrentParams } from "@/mandelbrot-state/mandelbrot-state";
 import { MandelbrotWorkerType, mandelbrotWorkerTypes } from "@/types";
 import BigNumber from "bignumber.js";
